[netcfg] Clean up test_track_socket_marks
Changes the test so that instead of being one detached background task
and a foreground task, it is two futures that are run in parallel using
futures::future::join.

This should solve a potential flake in which the shutdown channel is not
drained before the test finishes, causing a panic in the detached
background task. Now both tasks should wait to be completed before
destroying the test environment.

Multiply: policy-management-integration-test
Bug: 431384532
Change-Id: I3d4920c4f79b6088522154a6d41d48edb93b2820
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324845
Fuchsia-Auto-Submit: Zoey Bush <zmbush@google.com>
Commit-Queue: Zoey Bush <zmbush@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>