[archivist] Fix termination issue after flush
If the flush API was used, it's possible that termination would then not
work correctly. It would also impact the delay that the socket thread
has to reduce CPU impact.

Once a Zircon event fires with a user signal, before it can be reused,
it needs to be cleared, and it also needs to be reregistered with the
port.

This CL also includes some refactoring:

 1. I flattened the nested select loops in serial.rs.

 2. I changed some mpsc channels into one-shot channels.

Bug: 386831734
Change-Id: I710a9a43671dc51c53ac7585c41d5e2e4c370538
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324244
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>