[bt][rfcomm] Split the RX/TX datapaths and add backpressure
Split the RX and TX datapaths into separate async tasks which
continually perform async reads & writes to the relevant sockets.

In the TX side, the task will continuously read from the socket and
attempt to relay the packet to a centralized flow controller task. The
TX task will stop reading from the socket to signal backpressure to the
application until there are credits available.

In the RX side, the task will continuously read data frames processed by
the flow controller. If the application is slow to read the data frames,
the task's buffer will naturally fill up. The flow controller will use
this as a signal that the RX path is congested and stop issuing credits
to the remote device.

Update the FlowController trait to define a single run method. This acts
as the entrypoint for each of the controller's data processing task.

This should prevent dropping packets in high-throughput RFCOMM chanells
when there was sufficient lack of credits to transmit the frames, or
insufficient space in the receiving client to deliver them immediately.

Bug: 371346522
Test: fx test bt-rfcomm-tests
Test: Manual - OOBE locally many times
Test: Manual - Test Team custom build

Depends-on: turquoise-internal:I30d1defa5ee71486d89037ea13ae3a6f3e2cef7c

Change-Id: Ia913f09988ec398d3b24e28233863e128a9bdd1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1317645
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>