Revert "[logs] Actually add performance metrics"
This reverts commit 0edd7c5a0a9aa55d25aaaad4ebc4437ca9183ec9.

The test is flaky when we collect memory instrumentation, so we are
removing the instrumentation. The test will still run in non-perfcompare
configurations, but it will no longer collect memory profile data so
it's being removed from the perfcompare bots.

Reason for revert: See below bugs.

Bug: 432508629
Bug: 431831178

Bug: 411143842 Original change's description:
> [logs] Actually add performance metrics
>
> Actually add perf metrics.
>
> Bug: 411143842
> Change-Id: Ief464d8f7185430608594b18da6e7edbbd551744
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1256973
> Reviewed-by: Mark Seaborn <mseaborn@google.com>
> Commit-Queue: Brian Bosak <bbosak@google.com>

Bug: 411143842
Change-Id: I7e2dcfefb6efeac113a65c6811b3669f72c53e79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324553
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>