[kernel][phys] Stop accessing NVRAM through the physmap
Doing so complicates early bootstrap relating to PMM arena selection and
physmap construction: we don't want the NVRAM to be end up in a PMM
arena and we want to ensure that all non-arena ranges are
protected/unmapped in the physmap... with the exception of the NVRAM
range so long as we're accessing it through there. So we simply have
physboot start creating and handing off a dedicated mapping of the NVRAM
range, simplify the current physmap set-up by removing the NVRAM range
avoidance logic.

Tested manually on a Sorrel that crashlog generation still works with
this change.

Also realized in this change was that the NVRAM range was not being
added to the root resource filter. This may have been a longstanding
coverage, but in any case is fixed in this change.

Bug: 42164859, 430110290
Change-Id: I00cee8454aa0ae93af3cfa329857984729ea5135
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1303230
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>