[ffx] set_behavior() checks if it is already set
In set_behavior(), we should check to make sure it is not already
set. If it is, it's almost certainly a bug. (Note: even a "compatible"
behavior, such as from DirectConnector(_) to DirectConnector(_), is
still an indication of a bug, since each connector may contain data that
should be a global" singleton".)

Change-Id: I6a6a6964a973a2819a5e8fd44e2d5ab707fbbe66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324545
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>