[ffx] let ffx subtools declare #[target(direct)]
When `#[target(direct)]` is at placed at the top of an `FfxTool`
declaration, the tool will use a direct connection to the target rather
than going through the daemon. (This is in contrast to an upcoming
change which will give the user the ability to control whether a direct
connection is made. In this CL, the tool is declaring the requirement
for a direct connection.)

In order to avoid situations in which developers declare both
`#[no_target]` and `#[target(direct)]`, this CL also changes the
no_target declaration to `#[target(None)]`.

Change-Id: I6a6a6964c6bf51ccbe4e0488d4c842f254e0572d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324970
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>