[bt][hfp-hf] Transfer call to HF.
On FIDL request, request the AG start the SCO connection.  On a SCO
connection state change, produce a call state change reflecting tihs
from the Calls stream.

This CL also removes inserting new calls into the calls struct when an
outgoing call is made; instead the presence of a call is determined
entirely by AG call indicators.  This was needed in the case of a call
which starts on the AG and is later transferred to the HF.  Before it is
transferred, the only way the HF knows about it is via AG indicators, as
there is no FIDL API metohd creating the call. To make the logic
uniform, now all calls, including those initiated by a FIDL method to
the HF, are only inserted into the Calls struct when an indicator is
received from the AG.

Bug: b/421998483
Test: fx test bt-hfp-hands-free-tests
Change-Id: I0e272617d3ba275ee22fb45600272a7209fbd1d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1292774
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Matthew Kehrt <kehrt@google.com>