[fxfs] Add test for mark_dirty pages under mem pressure.
We now store the critical dirty bytes limit in a field in the
VolumesDirectory just to support changing it in tests, otherwise we'd
need to fill 1% of memory with dirty pages and that could get cumbersome
on some systems.

Also moves memory pressure injection into the TestFixture so that we can
easily set it for the whole filesystem and updates a bunch of tests to
stop setting up their own MemoryPressureMonitor.

Bug: 430394762
Change-Id: I6026126d78f87e7081c8af15a118e7464158c12c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324664
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>