[ffx][fuchsia-controller] Timeout on getting remote control proxy.
If we've verified there is an active connection to the device, and we're
not able to get a remote control proxy within the timeout, return an
error.

In addition, adds the proxy timeout to the config keys, and makes use of
it instead of a hard-coded value inside fuchsia-controller.

Bug: 430362260

Change-Id: Ida7c5de5ea131781226a7967a30bb4606a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1325246
Fuchsia-Auto-Submit: Andrew Davies <awdavies@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>