[kernel][vm] Remove offset_ from VmPageSpliceList
The 'offset_' in the VmPageSpliceList does not have proper logical
meaning, as a splice list is just a list, it does not exist anywhere
at any offset.

The only actual use for the offset is to initialize the skew in the
page_list_ when a splice list is populated by directly moving nodes
from an existing VmPageList. In this case the offset is known by the
TakePages call and can be used to initialize the skew, without needing
to otherwise record it.

Change-Id: If8c564853f29f4135c8a113a97041763d64ce57d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1322144
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>