[kernel][mutex] Move should_clear_ to policy state
The should_clear_ flag is inherently local to the acquisition, and not
the overall mutex state. This change therefore moves it into the
policies state field causing it to be allocated transiently on the
stack in the Guard objects, instead of permanently in the Mutex object
itself.

This does result in additional work for manual CriticalMutex
manipulations, as the should_clear state needs to be tracked by the
caller, but such usage is exceedingly uncommon as the Guard wrappers
are almost always used.

Benefit of this change is saving 8 bytes per CriticalMutex object.

Change-Id: I973770278a0bb4c90985940de19bd10df9155517
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1323351
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>