[fuchsia-pkg] Delete Package type.
This change deletes the Package struct, as a cleanup. To an outside
obvserver, it would seem the Package struct would be quite important,
but in fact it's pretty much an implementation detail of the package
build process.

This change inlines the little bit of functionality it had into build.rs
and package_manifest.rs.

Change-Id: I1d96a7d134b65e3231587773e267abf16aa0c928
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1322404
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Hunter Freyer <hjfreyer@google.com>