[wlan] Deauthenticate when SME says so.
The brcmfmac driver has generally used the C_DISASSOC command when
disconnecting from a BSS, even when SME has directed the driver to
deauthenticate. This CL changes the command used for SME-initiated
deauthentications to C_SCB_DEAUTHENTICATE_FOR_REASON. This command has
proven to be more effective in clearing firmware and AP connection
state, and is the correct command to use in this case: SME has
directed the driver to deauthenticate, and this command causes an
802.11 deauthenticate frame to be sent to the AP.

After this CL merges, there will be only one use of the C_DISASSOC
command left in brcmfmac: in the `brcmf_link_down` function, which calls
the C_DISASSOC command for a specific reason documented in the code.

This change also removes a few more instances of AP client count expects
from the brcmfmac sim framework wnm_test; these expects aren't a great
model of real world behavior in ambiguous cases where a real device may
or may not send a disconnect if a roam attempt fails midway.

Bug: 406330083, 369703077
Test: fx test bcrmfmac_test brcmfmac_sim_test;
      Antlion suites: beacon_loss_test, vape_interop_test,
      minimal.vim3-vg-release-antlion-test,
      smart_display_max_eng_arrested.sherlock-release-antlion-test-ns2-quick,
      wlan_target_security_test, wlan_phy_compliance_abg_test,
      ping_stress_test
Change-Id: I24fbfd99fdf85aa6251d5c963281e3a32b1028d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324464
Reviewed-by: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Karl Ward <karlward@google.com>