[wlan] Update channel configuration (chanspec).
When connecting or roaming, brcmfmac must send the appropriate channel
configuration (chanspec) to the firmware. Since the beginning of
brcmfmac, we have hard-coded the channel bandwidth to 20 MHz regardless
of the actual configuration (which can be larger than 20 MHz, e.g. 40
MHz). After this change, brcmfmac will use the bcmdhd-provided channel
configuration logic for most channel configurations (with the exception
of 80+80 bandwidth channel configuration).

The hard-coding of 20 MHz channel bandwidth is a long-standing issue
with the brcmfmac driver that has been implicated in some of the
failures where connect/roam attempts fail with a NO_NETWORK error. By
configuring the firmware with a correct channel configuration, we should
see a decrease in the incidence of NO_NETWORK failures, and that has
been the observed result in bench testing.

After this change, the old channel configuration logic will be used in
two places:

- bandwidth 80+80 configuration, which is difficult to test with current
  infra
- channel configuration for scanning, which is a separate code path

Bug: 42144507
Test: fx test brcmfmac_test brcmfmac_sim_test;
      Antlion suites beacon_loss_test, vape_interop_test,
      wlan_phy_compliance_11n_2g_ht40lower_open
Change-Id: If054ccee0ddc242818f169ec26b93a384a063ede
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1327144
Reviewed-by: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Karl Ward <karlward@google.com>