[trace] Refactor File I/O out of TraceTask
This removes the opening of the output file
from the TraceTask and replaces it with an
Async writer implermentor.

This allows the caller to open the file and
pass in the file object, and also allows on-device
usage of TraceTask to use a socket based TraceTask
and not need File I/O.

Additionally, it removes the dependency on async-fs
which is not available on Fuchsia.

Bug: 428951658

Change-Id: If6bd844d4c4fc5f40d5a1750bfd59c546ee5d868
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324714
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>