[netpol] Integration tests with FuchsiaNetworks
Update the DNS netemul integration test and add an integration test to
the netcfg integration test suite.

The DNS test used to interact directly with the FuchsiaNetworks protocol
to alter Socketproxy behavior. Now that netcfg manages the
FuchsiaNetworks connection, change these calls to a default route
add/removal to change whether the network is provided to the
Socketproxy. For simplicity, use NDP to reflect a DNS result from
netcfg's DNS perspective.

Add a test to netcfg to verify the socket marking behavior when a
Fuchsia network is present or absent. This test also validates a netcfg
variant where the Fuchsia network has delegated provisioning and should
not be used as a FuchsiaNetwork seen via the Socketproxy.

Bug: 385368910
Change-Id: I5c0ee14127322f5bbf8417b8d3b7c4894a10ecb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190185
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>