[archivist] Don't use the kernel log for logging
Change Archivist to log directly to itself for its own log
messages. Prior to  this change, Archivist would log to the kernel log
only to then read it back.

It would also redirect stdout and stderr to the kernel log. This is
problematic when the system is being shut down because writes to stdout
and stderr can fail, which result in a panic (which was the cause of
flakiness for some tests).

This CL also changes Archivist so that the embedded version of Archivist
also logs to itself, which removes the need for the log_to_debuglog
config option. This reduces the number of configurations that we need to
maintain.

NO_IFTTT=Not applicable

Bug: 396173225
Change-Id: I0b2cfd0946e599b93f6533d060bd71d2a0f2a2f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1323184
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Chris Suter <csuter@google.com>