[network][packet] Introduce LayoutBufferAlloc
Introduce an allocator trait that provides buffers with a given layout.
This allows serialize_new_buf to take a GrowBufferMut instead of a
ReusableBuffer, getting it more in line with the serialize method.

A buffer that only grows during serialization is easier to deal with
with the upcoming introduction of dynamic-dispatch serializers.

Bug: 428952721
Change-Id: I5f60a539cdd767d7a4359aec2b035808806877de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324945
Reviewed-by: Sergey Ulanov <sergeyu@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>