[trace-reader] Replace fbl::String with std::string
This change continues the effort to port the trace-reader library
from FBL to the C++ standard library by replacing fbl::String with
std::string.

In service of this goal, replace uses of fbl::StringBuffer with
std::string objects with reserved memory capacity and uses of
fbl::StringPrintf with std::stringstream.

Once C++20 is allowed in the SDK, std::format will be more drop-in
replacement for fbl::StringPrintf.

Bug: 432527736
Test: fx build //zircon/system/ulib/trace-engine/test:tests
Test: fx test host_x64/trace-reader-test-test
Test: fx test //src/performance/trace2json:trace2json_tests
Test: fx test //src/performance/lib/trace_converters:chromium_unittests

Change-Id: Ib902b8d41f8f1b0a1dc4018137684e5a33235172
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1325217
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>