[starnix][kernel] Use connection rights for POSIX mode bits
When synthesizing POSIX mode bits, instead of querying the node's
underlying abilities, we now use the permissions used when the node is
actually created. This better reflects what operations are actually
possible on that node within Starnix itself, as this will be limited by
the flags/rights used to open the connection.

Rework POSIX mode type mapping for `new_remote_file` so that we avoid
using `zxio_get_posix_mode` which is in the process of being removed.

Test: fx test starnix-tests
Change-Id: I96d29d0f9528a8a12a120761143d037fafa366d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1327205
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>