[fshost] export the PartitionService for the right gpt on gpt_all
Add a new matcher that gets added for gpt all, instead of always adding
the system gpt matcher and possibly a followup. This allows the explicit
matcher to peek at the partitions in the gpt to decide if it wants to
tag this particular gpt instance as the system gpt.

This heuristic is copied from the paver, looking for a partition named
"super" or "fvm". Ultimately the paver is the only thing that cares
about what is hooked up to this so it makes sense to move the heuristic
to fshost when we can only forward one service.

Fixed: 428765340
Change-Id: I068895ce72b98faea4196ca1cb122e9205bab4fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1323924
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>