[kernel][pmm] Introduce WaitForSinglePageAllocation
Replace usages of `WaitTillShouldRetrySingleAlloc` and their retry
patterns for this method.

The new mechanism, relies on an AutoUnsignalEvent for page_free_evt,
which will only wake a single thread at a time, and once such thread
performs a successful allocation, signals the following thread.

Spurious wake up may occur as we race against other parts of the system
to perform allocation, in which case it is ok to go back to sleep, and
wait for the event to become signalled again.

Additionally, in order to properly enforce a panic when the systems
seems NOT to make progress over a period of time, when PmmNode signals
that memory is available, we reset the retry count.

Change-Id: I078117f6257c7ac70980f44ffa6b5fad4edf7f4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1321599
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>