[kernel] uart adds regions to deny list
Before we had platform specific hook to add a UART's region(pio or mmio)
to the resource deny list. Now that `root_resource_filter` is the one
mechanism to deny access to resources, we no longer need these hooks.

Introduced `ioports` for PioDrivers on as an accessor equivalent to
`mmio_range` for MmioDrivers.

Change-Id: I8ad0d0396f2ce3dca8736df3329ed29cc239dc3b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1316306
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>