[scenic] Store the display mode in Display.
The Mode struct obtained from the display stack needs to be passed as-is
in SetDisplayMode(). So, it makes sense to store it, rather than store
most of its members as individual data members.

This prefactoring facilitates the fix for the attached bug.

Bug: 402804098
Test: fx test //src/ui/scenic:display_unittests
Test: fx test //src/ui/scenic:flatland_unittests
Test: fx test //src/ui/scenic:flatland_engine_unittests
Change-Id: Ieb7c9283a790d15466b0613d87efd92465df0c91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1328369
Reviewed-by: Emircan Uysaler <emircan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>