[starnix] Store overridden creds on CurrentTask
In a number of situations (acting on behalf of outside-Starnix
entities, overlayfs, ...), the kernel needs to temporarily switch
credentials to perform action as if the currently running task had
different credentials.

It is not safe to simply change the credentials of the task, as those
are visible to the outside world and used for access checks where the
task is a target. Instead, we introduce a set of optional overridden
credentials on CurrentTask. These must be used instead of the Task's
real_creds when the CurrentTask is acting.

This temporarily duplicates the SEStarnix effective_sid mechanism. The
next CL will replace the SEStarnix effective_sid with using overridden
creds directly.

We put these creds behind a RefCell. Ideally, this wouldn't be needed,
as the overridden credentials are only ever accessed and mutated on
the CurrentTask, but CurrentTask is currently passed as an immutable
reference everywhere, and we'd have to turn the reference mutable,
which is a pretty large refactoring to take on.

Change-Id: Id6c9bf059678eaafc0ce3778144dc440eeece4e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1316024
Commit-Queue: Ambre Williams <ambre@google.com>
Reviewed-by: Wez <wez@google.com>
Fuchsia-Auto-Submit: Ambre Williams <ambre@google.com>