[kernel][vm] Remove is_any_borrowing_enabled
As per the also removed code comment, this was added as a way to ensure
borrowing sites did not fail to check their per site config for whether
or not they should be borrowing. If there were to be such a discrepancy
then we should seek to identify it robustly with tests and other
validation, not allow the system to silently continue in a buggy state.
As far as is known all borrowing sites are correctly guarded by their
respective config checks.

Change-Id: I5abf6fe55103b16ad885d35373f21e227fb1703e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1323024
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>