[starnix] Use current creds for temporary security state changes
Now that we have introduced a "current credentials" concept, we can
use the current security state to host temporary security state
changes instead of using the real, externally-visible security state.

As a consequence, we retire the TaskAttrsOverride mechanism in favor
of directly changing the current credentials. We remove the
effective_sid task attribute, and as a consequence the
task_effective_sid helper function, since this is now "just" the
current_sid of the current task (so the "effective" name only brings
confusion). This change shows a number of places where we forgot to
use the effective_sid where it would have been appropriate - we hope
the new mechanism will make it easier to get right by forcing to spell
out "real" or "current" once we put the security state next to the
credentials.

Bug: 413607595

Change-Id: Iff085e7c60659bf7cc5343080baa23a3b7395d33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1327004
Reviewed-by: Wez <wez@google.com>
Fuchsia-Auto-Submit: Ambre Williams <ambre@google.com>
Commit-Queue: Ambre Williams <ambre@google.com>