[kernel][x86] Single ConsistencyManager
Uses a single consistency manager in x86 page table updates. A separate
consistency manager does not need to be constructed for the
referenced_pt_ updates, as the existing |cm| will already forward TLB
invalidations to the referenced_pt_. The only caveat is that the
invalidations must be done before the lock is dropped.

In practice adding missing page tables does not actually require any
TLB invalidations, so the flush will be a no-op, and for the
RemoveMapping case by re-ordering the referenced_pt_ to be updated
second, the Flush is likely to be the last action before the overall
operation terminates, allowing the flush in the top level Unmap
caller to be a no-op.

Removing the separate consistency manager means that the space for one
does not need to be reserved on the stack for every recursive call of
MapPages, of which there are 4. As the consistency manager is quite a
large object, 312 bytes this reduces stack usage by 1248 bytes in this
path.

Change-Id: Ic206548db4045885a5f395d82b043268b984566b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1314324
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>