[kernel][psci][arm64] Enable restriction of CPU_SUSPEND power state
platform_suspend_cpu now allows the caller to specify whether more
than just the calling CPU is allowed to be powered down.  We now
choose the best matching power down state for the operation.

If domain power down is allowed, and the calling CPU is CPU-0, and the
PSCI implementation provides a power domain power state, then it will
be used.  Otherwise the CPU-only power state will be used.

This change ensures that platform_suspend_cpu_tests work on Sorrel
even when the calling CPU happens to be CPU-0.

Later on, when implement "deep idle" we can use this same
platform_suspend_cpu argument to ensure we don't inadvertently power
down the domain when the boot CPU happens to go idle while the
secondaries are already in idle.

Bug: 417558115
Change-Id: I4814f0eaa58aa1f465fe7fee18d28e0139219e14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324844
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>