[build][idk] Check categories and API areas in IdkGenerator
Use sdk_common to check atoms in a collection for category and area
violations. Atom ID and invalid types are also checked.

Only atoms in the collection are checked, so this is not yet a full
replacement for the checks of every `sdk_atom()` that are performed by
create_atom_manifest.py create_molecule_manifest.py.

Also add "api_area", when specified, to prebuild info in order to
support these checks.

Bug: 419105478
Test: Tests added to `fx test --host //build/sdk/generate_prebuild_idk:idk_generator_unittest`
Change-Id: I51ee356504285fab2384294f12c9957fed411972
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1313807
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>