[vfs][cpp] Avoid reporting mode bits where not required
The use of the POSIX mode bit field is intended for use in mutable
filesystems that support POSIX attributes. For all other use cases, we
use fuchsia.io-specific functionality to determine the type or other
properties of a node (e.g. using fuchsia.io/Node.GetAttributes, or via
zxio_attr_get). We still synthesize mode bits at the VFS level for
backwards compatibility with fuchsia.io/Node.DeprecatedGetAttr, but this
will be cleaned up when we remove support for that.

Change-Id: Ib8395df843b739d2b55df61cdca666adae87bf56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1327385
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>