[component_manager] Consume the namespace Dict
Once the namespace Dict is converted to a vfs, there is no need to keep
the original Dict around.

The original API nominally consumes the Dict, but since Dict wraps an
Arc it really just consumed the reference, it was not consuming the
contents.

Bug: 419024317
Change-Id: I19d341761896661eb07bc58ab494edc953a6e91c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1328366
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Gary Bressler <geb@google.com>