[zxdump] Handle 32-bit ELF images in <lib/zxdump/elf-search.h> API
This extends the zxdump::DetectElf and zxdump::DetectElfIdentity
functions to handle Elf32<> and Elf64<> formats, their signatures
now using variant types for the phdrs.

This is tested via ld::RemoteDynamicLinker to add both 64-bit and
32-bit images to the same address space.

The armv7-linux-gnueabihf test images happened to uncover an
unrelated existing bug in the dumper code related to ElfSearch
but not in the ElfSearch API itself.  That results in double
reporting of ELF files where the first file page is mapped in
multiple segments.  The ElfSearch API handles this correctly but
the NT_GNU_BUILD_ID note generation in core dumps does not.  This
will be fixed in a later change, which will may require changes
to the zxdump::SegmentCallback API contract as well as a deeper
rework of ProcessDump::FindBuildIdNote and its users.

Bug: 395606674
Change-Id: Ifdb656cfae29ecb9f75f655bf9f258b2ad17416b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1300124
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>