[fxfs] Blob overwrite doesn't return until tombstoning.
Moving the open count sub one future as a return to be run outside of
the sync context of the commit callback, rather than spawning the task.
This ensures that the tombstoning is done before returning so that in
the tests we can flush the graveyard to ensure that the old version is
actually deleted.

Bug: 430394762
Change-Id: I031a9c84df28538f0d43461ae5baae5086be39bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1329224
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>