[firmware] fix fastboot lacewing flake
Occasionally lacewing will return from `boot_to_fastboot_mode()` but
trying to talk to the device will still give a brief "waiting for
<serial>" message, which messes up our `getvar` parsing logic.

We had previously tried to solve this by querying a variable after
booting to fastboot so we know the device can communicate, but this
doesn't seem to be sufficient. It's not clear why after returning from
`getvar` it might still say "waiting for ..." in subsequent commands,
but it does.

Regardless of cause, this seems pretty harmless, so this attempt allows
us to ignore the "waiting for" line on any individual `getvar` query.

Bug: b/419262916
Change-Id: I5c7c8743e83abecc9777ddf96b485c3c7dbbadba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1329430
Fuchsia-Auto-Submit: David Pursell <dpursell@google.com>
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>