[socket-proxy] Add component to reboot policy
Socket-proxy should be added to the reboot_on_terminate security policy
so that the device can restart when the component unexpectedly stops.

Context for this change is in b/417568023. Socket-proxy will become the
component that is vending sockets for the system. FDIO caches the socket
provider and does not attempt to reconnect, so if the socket-proxy
restarts then new sockets will be unable to be created anyway. Fixing
this in FDIO is tracked in the associated bug.

Confirmed via `ffx component stop` on a socketproxy-enabled product
that the device shuts down.

This change to make the component critical landed in
https://fxrev.dev/1276064.

Bug: 417568023
Change-Id: I2e1a6c848e44886796810c27fcb639803119521d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1329389
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>