[kernel][vm] MarkMergeable takes RefPtr by value
Passing by rvalue was not necessary, as MarkMergeable does not actually
consume the RefPtr, and failure to consume the RefPtr meant it remained
with the caller. The tests using MarkMergeable both used ktl::move, yet
then dereferenced the ptr that they expected to be moved, which worked
due to MarkMergeable not actually consuming the RefPtr like it claimed.

This commit therefore restructures the test to not reference the
mappings after they are moved.

Change-Id: I5f61812f4150637959158b609823615b877cbedd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1323345
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>