[fxt] Add blob argument type
This adds a blob argument type (10) to the FXT format. It updates the
documentation and various implementations and tools that deal with the
FXT format in-tree.

Blob arguments contain opaque streams of octets. Their size is limited
by the maximum size of arguments - 32kB, but that size includes an 8
byte header and optionally an inline name string. The original size of
the blob contents is stored in the argument header so that a reader can
both ignore padding bytes and determine if the contents have been
truncated.

Bug: 433532694
Change-Id: I270910ac7ac6b789a7958e4d268ecd3f59451870
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1277747
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
API-Review: Hunter Freyer <hjfreyer@google.com>