[netlink] Make unmanaged tables unopinionated
This is to prepare for the support of interface-local route tables.
In today's world, the only unmanaged table is the main table and the
existing code overfits the assumption. For example an unmanaged table
is not removable; the `RouteTableKey` assumes all non-main tables as
managed. These do not work with the interface-local route tables as
those are unmanaged but can be removed. This change tries to simplify
by removing the assumption for unmanaged table. The `RouteTableMap`
is just acting as a map between netlink indices and fidl tables and
does not handle differently for the main index.

Bug: 427226662
Change-Id: I6a6a6964ddece7b944ae2cf89ccaaf86e645bb87
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1329160
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>