[starnix] Lock task command line and credentials separately
The credentials are now under a RwLock. The lock is never held for
long, so we do not need static lock ordering annotations yet. However,
a future CL will ensure that we keep the credential locked long enough
to avoid TOCTOU issues, and we will introduce a specific lock level
before doing that.

Bug: 433463756

Change-Id: Ifae61a04420fb70b6091195249384b63b04533e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1322564
Commit-Queue: Ambre Williams <ambre@google.com>
Reviewed-by: Wez <wez@google.com>