[starnix] Statically link zxio into syncio
This makes the syncio library used by starnix to statically link the
zxio library instead of dynamically linking it. This allows optimizing
calls through this library more precisely.

Since zxio uses the C++ standard library, this means that binaries using
the library need to supply a C++ standard library implementation. The
zxio_static target adds a public configuration to statically link libc++
which applies to C++ and Rust targets. Intermediate libraries depending
on syncio need to be careful to declare this dependency a public dep in
order to propagate this configuration to the terminal binary.

Change-Id: Id304ebf60004cb126c836f8353c056e5cebd7757
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233108
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Adam Barth <abarth@google.com>