[fxfs] Move scope guards out of PagerRange and rely on the scope.
The invariant that we try to preserve in the tests are that references
to the volumes do not outlive the fixture, so we try to verify that
there are no more references during `fixture.close()`. So during volume
termination we wait on execution scope before returning to ensure that
all references in requests are done. Files always have a reference back
to their owing volume via their handles, so leftover files can cause
this as well.

With the scope guard being in PagerRange it is not possible to ensure
that the file references go away first because there is always a file
reference outside the PagerRange which needs to outlive the PagerRange
so that the transfer buffer can be borrowed during `supply_pages`.

Bug: 430394762
Change-Id: I507660b15e369a2f7d9bbcdfcb68745da1af1ec5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1330284
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>